# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Vault API
  description: >
    Osobní hudební systém — REST API. Pokrývá globální katalog/vyhledávání
    (MusicBrainz/Deezer proxy), asynchronní on-demand provisioning chybějících
    skladeb a doporučení/playlisty (ListenBrainz). Real-time synchronizace
    přehrávání je popsaná odděleně v `asyncapi.yaml`.
  version: "0.1.0"
servers:
  - url: https://vault.local/api/v1
    description: Přes reverse proxy / VPN (Tailscale), nikdy veřejně
security:
  - bearerAuth: []
...
paths:
...
  /recommendations/trending:
    get:
      tags: [recommendations]
      summary: >
        Nejhranější nahrávky napříč veřejnou komunitou ListenBrainz
        (https://api.listenbrainz.org/1/stats/sitewide/recordings) — ne data
        z téhle instance, server sám žádné poslechy nesleduje.
      parameters:
        - name: range
          in: query
          schema: { type: string, enum: [week, month, year, all_time], default: week }
        - name: limit
          in: query
          schema: { type: integer, default: 20, minimum: 1, maximum: 50 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Recording" }
  /recommendations/my-top:
    get:
      tags: [recommendations]
      summary: >
        Nejposlouchanější nahrávky configurovaného účtu (LISTENBRAINZ_USERNAME)
        přímo z jeho ListenBrainz statistik — na rozdíl od /daily-jams funguje
        i bez vygenerovaného troi doporučení, stačí mít zaznamenané poslechy.
      parameters:
        - name: range
          in: query
          schema: { type: string, enum: [week, month, year, all_time], default: month }
        - name: limit
          in: query
          schema: { type: integer, default: 20, minimum: 1, maximum: 50 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Recording" }
  /recommendations/community:
    get:
      tags: [recommendations]
      summary: >
        Doporučení odvozená z poslechů uživatelů s podobným vkusem na veřejném
        ListenBrainz (similar-users + jejich top nahrávky). Prázdné, dokud
        LISTENBRAINZ_USERNAME není skutečný účet na listenbrainz.org s
        poslechovou historií.
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 20, minimum: 1, maximum: 50 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Recording" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Device-scoped JWT vydaný při párování zařízení.
...
  schemas:
...
    Availability:
      type: string
      enum: [available, provisionable, unavailable]
      description: >
        available = MediaAsset.status==AVAILABLE; provisionable = existuje
        v katalogu a lze založit provisioning job; unavailable = žádný
        známý zdroj (čistě metadatová položka bez cesty k obstarání).
...
    Recording:
      type: object
      properties:
        id: { type: string, format: uuid }
        mbid: { type: string, nullable: true }
        releaseId: { type: string, format: uuid, nullable: true }
        artistId: { type: string, format: uuid }
        title: { type: string }
        durationMs: { type: integer }
        isrc: { type: string, nullable: true }
        trackNumber: { type: integer, nullable: true }
        availability: { $ref: "#/components/schemas/Availability" }
        previewUrl:
          type: string
          format: uri
          nullable: true
          description: 30s náhled z Deezer (doplňkové metadata, best-effort — nemusí být vyplněno).
        listenCount:
          type: integer
          nullable: true
          description: >
            Počet poslechů z veřejného ListenBrainz API (doplňkové metadata,
            jen u /recommendations/trending a /recommendations/community —
            jinde vždy null).
...